Time: 0 ms (100%), Space: 7.7 MB (98.8%) - LeetHub

+++ 0070-climbing-stairs/0070-climbing-stairs.cpp
@@ -5,9 +5,9 @@ public:
             return n;
         }
 
-        long previous1 = 3;
-        long previous2 = 2;
-        long sum;
+        long long previous1 = 3;
+        long long previous2 = 2;
+        long long sum;
 
         for (int i = 4; i <= n; ++i) {
             sum = previous1 + previous2;
